Set up braille_translator class and tests
